requirements.txt
mediapipe
diffusers
torchvision
opencv-python
streamlit
gradio
tqdm
transformers